requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
backports.zoneinfo==0.2.1
bardapi==0.1.11
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
chromadb==0.3.26
click==8.1.3
clickhouse-connect==0.6.1
colorama==0.4.6
coloredlogs==15.0.1
crispy-bootstrap5==0.7
cryptography==41.0.1
dataclasses-json==0.5.7
Django==4.2.1
django-crispy-forms==2.0
django-restframework==0.0.1
djangorestframework==3.14.0
docx2txt==0.8
duckdb==0.8.0
elastic-transport==8.4.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fastapi==0.96.0
filelock==3.12.0
flatbuffers==23.5.26
frozenlist==1.3.3
fsspec==2023.5.0
graphlib-backport==1.0.3
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
httptools==0.5.0
huggingface-hub==0.15.1
humanfriendly==10.0
idna==3.4
langchain==0.0.189
lxml==4.9.2
lz4==4.3.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
onnxruntime==1.15.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==9.5.0
posthog==3.0.1
protobuf==4.23.2
pulsar-client==3.2.0
pycparser==2.21
pydantic==1.10.8
PyMuPDF==1.22.3
pypdf==3.9.1
pypdfium2==4.14.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
sqlparse==0.4.4
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.22.0
Wand==0.6.11
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
zstandard==0.21.0
